backend/python-ml/requirement.txt
flask
scikit-learn
requests
pandas
python-dotenv
flask-cors
gunicorn